requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9
qrcode[pil]==7.4.2
pytest==8.3.0
pytest-cov==5.0.0
